cyna-security-pipeline/requirements.txt
# Dépendances du service pipeline (ingestion + enrichissement + Security-Log-Generator)
psycopg2-binary==2.9.9
python-dotenv==1.0.1
requests==2.31.0
PyYAML==6.0.1
Faker==16.3.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.7.5
